--- sandbox/video-conferencing-ms-example/.env.example
@@ -1,11 +1,0 @@
-VONAGE_API_KEY=
-VONAGE_API_SECRET=
-TIME_TO_START=
-DB_HOST=
-DB_PORT=
-DB_USER=
-DB_PASSWORD=
-DB_DATABASE=
-DB_SCHEMA=
-JWT_SECRET=
-JWT_ISSUER=
