--- services/video-conferencing-service/.env.example
@@ -1,8 +1,0 @@
-DB_HOST=
-DB_PORT=
-DB_USER=
-DB_PASSWORD=
-DB_DATABASE=
-DB_SCHEMA=
-VONAGE_API_KEY=
-VONAGE_API_SECRET=
