+++ services/video-conferencing-service/src/controllers/video-chat-archive.controller.ts
@@ -1,5 +1,5 @@
 import {inject} from '@loopback/core';
-import {del, get, param, HttpErrors, put, requestBody} from '@loopback/rest';
+import {del, get, param, put, requestBody} from '@loopback/rest';
 import {authenticate, STRATEGY} from 'loopback4-authentication';
 import {authorize} from 'loopback4-authorization';
 import {PermissionKeys} from '../enums/permission-keys.enum';
@@ -14,22 +14,20 @@ import {
   CONTENT_TYPE,
   OPERATION_SECURITY_SPEC,
 } from '@sourceloop/core';
-import {repository} from '@loopback/repository';
-import {VideoChatSessionRepository, AuditLogsRepository} from '../repositories';
-import moment from 'moment';
+
+import { service } from '@loopback/core';
+import { ChatArchiveService } from '../services/chatArchive.service';
 
 export class VideoChatArchiveController {
   constructor(
-    @inject(VideoChatBindings.VideoChatProvider)
-    private readonly videoChatProvider: VideoChatInterface,
-    @repository(VideoChatSessionRepository)
-    private readonly videoChatSessionRepository: VideoChatSessionRepository,
-    @repository(AuditLogsRepository)
-    private readonly auditLogRepository: AuditLogsRepository,
-  ) {}
+    @service(ChatArchiveService) public chatArchiveService: ChatArchiveService,
+
+  
+   
+  ) { }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.GetArchives]})
+  @authorize({ permissions: [PermissionKeys.GetArchives] })
   @get('/archives/{archiveId}', {
     description:
       'Used to fetch a specific archive w.r.t archiveId. If archive is not present, it will throw HTTP Not Found Error.',
@@ -47,27 +45,11 @@ export class VideoChatArchiveController {
     },
   })
   async getArchive(@param.path.string('archiveId') archiveId: string) {
-    const archiveExists = await this.videoChatSessionRepository.findOne({
-      where: {
-        archiveId: archiveId,
-      },
-    });
-    if (!archiveExists) {
-      const errorMessage = 'Archive Not Found';
-      await this.auditLogRepository.create({
-        action: 'archive',
-        actionType: 'getArchive',
-        before: {archiveId},
-        after: {errorMessage: errorMessage},
-        actedAt: moment().format(),
-      });
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-    return this.videoChatProvider.getArchives(archiveId);
+    return this.chatArchiveService.getArchive(archiveId);
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.GetArchives]})
+  @authorize({ permissions: [PermissionKeys.GetArchives] })
   @get('/archives', {
     description:
       'Used to fetch a list of archives (meetings that were recorded).',
@@ -85,11 +67,11 @@ export class VideoChatArchiveController {
     },
   })
   async getArchives() {
-    return this.videoChatProvider.getArchives(null);
+    return this.chatArchiveService.getArchives();
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.DeleteArchive]})
+  @authorize({ permissions: [PermissionKeys.DeleteArchive] })
   @del('/archives/{archiveId}', {
     description:
       'Used to delete a specific archive w.r.t archiveId. If archive is not present, it will throw HTTP Not Found Error.',
@@ -109,27 +91,11 @@ export class VideoChatArchiveController {
   async deleteArchive(
     @param.path.string('archiveId') archiveId: string,
   ): Promise<void> {
-    const archiveExists = await this.videoChatSessionRepository.findOne({
-      where: {
-        archiveId: archiveId,
-      },
-    });
-    if (!archiveExists) {
-      const errorMessage = 'Archive Not Found';
-      await this.auditLogRepository.create({
-        action: 'archive',
-        actionType: 'getArchive',
-        before: {archiveId},
-        after: {errorMessage: errorMessage},
-        actedAt: moment().format(),
-      });
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-    return this.videoChatProvider.deleteArchive(archiveId);
+    return this.chatArchiveService.deleteArchive(archiveId);
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.SetUploadTarget]})
+  @authorize({ permissions: [PermissionKeys.SetUploadTarget] })
   @put('/archives/storage-target', {
     description:
       'Configures custom storage target to a custom Amazon s3 bucket or Microsoft Azure Storage.',
@@ -149,35 +115,6 @@ export class VideoChatArchiveController {
   async setUploadTarget(
     @requestBody() body: S3TargetOptions | AzureTargetOptions,
   ): Promise<void> {
-    const {accessKey, secretKey, bucket} = body as S3TargetOptions;
-    const {accountName, accountKey, container} = body as AzureTargetOptions;
-    if (
-      !(accessKey && secretKey && bucket) &&
-      !(accountName && accountKey && container)
-    ) {
-      throw new HttpErrors.BadRequest(
-        'Missing s3/azure credentials. Please check request body',
-      );
-    }
-    await this.videoChatProvider.setUploadTarget(body);
+    return this.chatArchiveService.setUploadTarget(body);
   }
-  // TO-DO: will do modifications later
-  // @authenticate(STRATEGY.BEARER)
-  // @authorize({permissions:[PermissionKeys.SetUploadTarget]})
-  // @del('/archives/storage-target', {
-  //   responses: {
-  //     [STATUS_CODE.OK]: {
-  //       content: {
-  //         [CONTENT_TYPE.TEXT]: {
-  //           schema: {
-  //             type: 'text',
-  //           },
-  //         },
-  //       },
-  //     },
-  //   },
-  // })
-  // async deleteCustomStorageTarget() {
-  //   return this.videoChatProvider.deleteUploadTarget();
-  // }
 }
