+++ services/video-conferencing-service/src/controllers/video-chat-session.controller.ts
@@ -1,11 +1,10 @@
-import {inject} from '@loopback/context';
-import {repository} from '@loopback/repository';
+
 import {
   param,
   patch,
   post,
   requestBody,
-  HttpErrors,
+  
   get,
   getModelSchemaRef,
 } from '@loopback/rest';
@@ -13,10 +12,10 @@ import {authorize} from 'loopback4-authorization';
 import {
   MeetingOptions,
   SessionOptions,
-  VideoChatInterface,
+
   SessionResponse,
 } from '../types';
-import {VideoChatBindings} from '../keys';
+
 import {authenticate, STRATEGY} from 'loopback4-authentication';
 import {PermissionKeys} from '../enums/permission-keys.enum';
 import {
@@ -24,34 +23,23 @@ import {
   CONTENT_TYPE,
   OPERATION_SECURITY_SPEC,
 } from '@sourceloop/core';
-import moment from 'moment';
-import cryptoRandomString from 'crypto-random-string';
+
 import {VideoChatSession, SessionAttendees} from '../models';
-import {
-  AuditLogsRepository,
-  VideoChatSessionRepository,
-  SessionAttendeesRepository,
-} from '../repositories';
-import {VonageBindings} from '../providers/vonage/keys';
-import {VonageConfig, VonageSessionWebhookPayload} from '../providers/vonage';
-import {VonageEnums} from '../enums';
+
+import { VonageSessionWebhookPayload} from '../providers/vonage';
+
+import { service } from '@loopback/core';
+import { ChatSessionService } from '../services/chatSession.service';
 
 export class VideoChatSessionController {
   constructor(
-    @repository(VideoChatSessionRepository)
-    private readonly videoChatSessionRepository: VideoChatSessionRepository,
-    @inject(VideoChatBindings.VideoChatProvider)
-    private readonly videoChatProvider: VideoChatInterface,
-    @repository(AuditLogsRepository)
-    private readonly auditLogRepository: AuditLogsRepository,
-    @repository(SessionAttendeesRepository)
-    private readonly sessionAttendeesRepository: SessionAttendeesRepository,
-    @inject(VonageBindings.config, {optional: true})
-    private readonly config: VonageConfig,
-  ) {}
+    @service(ChatSessionService) public chatSessionService: ChatSessionService,
+   
+   
+  ) { }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.CreateSession]})
+  @authorize({ permissions: [PermissionKeys.CreateSession] })
   @post('/session', {
     description: `Used for Creating a session with options such as end to end encryption, archive mode. 
       Note: Archiving Option cannot be enabled while using end to end encryption, otherwise 
@@ -61,7 +49,7 @@ export class VideoChatSessionController {
     responses: {
       [STATUS_CODE.OK]: {
         content: {
-          [CONTENT_TYPE.TEXT]: {schema: {type: 'string'}},
+          [CONTENT_TYPE.TEXT]: { schema: { type: 'string' } },
         },
       },
     },
@@ -70,58 +58,11 @@ export class VideoChatSessionController {
     @requestBody()
     meetingOptions: MeetingOptions,
   ): Promise<string> {
-    let scheduledTime: Date = new Date();
-    const auditLogPayload = {
-      action: 'session',
-      actionType: 'meeting-link',
-      actedAt: moment().format(),
-      before: meetingOptions,
-      after: {},
-    };
-    let errorMessage: string;
-
-    if (meetingOptions.isScheduled) {
-      if (!meetingOptions.scheduleTime) {
-        errorMessage = 'Schedule time is not set.';
-        auditLogPayload.after = {errorMessage};
-        await this.auditLogRepository.create(auditLogPayload);
-        throw new HttpErrors.BadRequest(errorMessage);
-      } else if (isNaN(meetingOptions.scheduleTime.valueOf())) {
-        errorMessage = 'Scheduled time is not in correct format.';
-        auditLogPayload.after = {errorMessage};
-        await this.auditLogRepository.create(auditLogPayload);
-        throw new HttpErrors.BadRequest(errorMessage);
-      } else if (moment().isAfter(meetingOptions.scheduleTime)) {
-        errorMessage = `Meeting can't be scheduled with schedule time in past!`;
-        auditLogPayload.after = {errorMessage};
-        await this.auditLogRepository.create(auditLogPayload);
-        throw new HttpErrors.BadRequest(errorMessage);
-      } else {
-        scheduledTime = meetingOptions.scheduleTime;
-      }
-    }
-
-    const meetingResp = await this.videoChatProvider.getMeetingLink(
-      meetingOptions,
-    );
-
-    const meetingLinkId = cryptoRandomString({length: 10, type: 'url-safe'});
-    const videoSessionDetail = new VideoChatSession({
-      sessionId: meetingResp.sessionId,
-      meetingLink: meetingLinkId,
-      isScheduled: meetingOptions.isScheduled,
-      scheduleTime: new Date(scheduledTime),
-      isArchived: meetingResp.isArchived,
-    });
-
-    await this.videoChatSessionRepository.save(videoSessionDetail);
-    auditLogPayload.after = videoSessionDetail;
-    await this.auditLogRepository.create(auditLogPayload);
-    return meetingLinkId;
+    return this.chatSessionService.getMeetingLink(meetingOptions);
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.GenerateToken]})
+  @authorize({ permissions: [PermissionKeys.GenerateToken] })
   @post('/session/{meetingLinkId}/token', {
     description: `Used for Generating token, which is used for connecting to a room/session on a client side. 
       In vonage, there are three different roles (Moderator, Subscriber, Publisher). 
@@ -146,86 +87,12 @@ export class VideoChatSessionController {
     sessionOptions: SessionOptions,
     @param.path.string('meetingLinkId') meetingLinkId: string,
   ): Promise<SessionResponse> {
-    const auditLogPayload = {
-      action: 'session',
-      actionType: 'get-token',
-      before: {meetingLinkId},
-      actedAt: moment().format(),
-      after: {},
-    };
-    let errorMessage;
 
-    if (typeof meetingLinkId !== 'string' || !meetingLinkId) {
-      errorMessage = 'Meeting link should be a valid string';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    if (
-      sessionOptions.expireTime &&
-      isNaN(sessionOptions.expireTime?.valueOf())
-    ) {
-      errorMessage = 'Expire time is not in correct format.';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    if (moment().isAfter(sessionOptions.expireTime)) {
-      errorMessage = 'Expire time can not be in past.';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    const session = await this.videoChatSessionRepository.findOne({
-      where: {
-        meetingLink: meetingLinkId,
-      },
-    });
-
-    if (!session) {
-      errorMessage = 'Session does not exist';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    // check if meeting is already ended
-    if (session.endTime) {
-      errorMessage = 'This meeting has been expired';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    // check for schduled meeting:
-    if (session.isScheduled && session.scheduleTime) {
-      if (
-        moment()
-          .add(this.config.timeToStart, 'minutes')
-          .isBefore(session.scheduleTime)
-      ) {
-        errorMessage = `Meeting can only be started ${this.config.timeToStart} minutes before
-         the scheduled time`;
-        auditLogPayload.after = {errorMessage};
-        await this.auditLogRepository.create(auditLogPayload);
-        throw new HttpErrors.BadRequest(errorMessage);
-      }
-    }
-
-    if (!session.startTime) {
-      await this.videoChatSessionRepository.updateById(session.id, {
-        startTime: new Date(),
-      });
-    }
-
-    return this.videoChatProvider.getToken(session.sessionId, sessionOptions);
+    return this.chatSessionService.getMeetingToken(sessionOptions, meetingLinkId)
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.EditMeeting]})
+  @authorize({ permissions: [PermissionKeys.EditMeeting] })
   @patch('/session/{meetingLinkId}', {
     description: 'Used for editing the meeting',
     security: OPERATION_SECURITY_SPEC,
@@ -240,86 +107,17 @@ export class VideoChatSessionController {
     @requestBody({
       content: {
         [CONTENT_TYPE.JSON]: {
-          schema: getModelSchemaRef(VideoChatSession, {partial: true}),
+          schema: getModelSchemaRef(VideoChatSession, { partial: true }),
         },
       },
     })
     body: Partial<VideoChatSession>,
   ): Promise<void> {
-    const {isScheduled, scheduleTime} = body;
-    let errorMessage = '';
-
-    const auditLogPayload = {
-      action: 'session',
-      actionType: 'edit-meeting',
-      before: {meetingLinkId, isScheduled, scheduleTime},
-      actedAt: moment().format(),
-      after: {},
-    };
-
-    const sessionDetail = await this.videoChatSessionRepository.findOne({
-      where: {
-        meetingLink: meetingLinkId,
-      },
-    });
-
-    if (!sessionDetail) {
-      errorMessage = `Meeting link ${meetingLinkId} not found`;
-      auditLogPayload.after = {
-        errorMessage,
-      };
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-
-    if (isScheduled && !scheduleTime) {
-      errorMessage = `Schedule Time is required if isScheduled is set to true`;
-      auditLogPayload.after = {
-        errorMessage,
-      };
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    if (scheduleTime && isNaN(moment(scheduleTime).valueOf())) {
-      errorMessage = `Schedule Time is Not in correct format`;
-      auditLogPayload.after = {
-        errorMessage,
-      };
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    if (moment().isAfter(scheduleTime)) {
-      errorMessage = 'Schedule Time cannot be set in the past';
-      auditLogPayload.after = {
-        errorMessage,
-      };
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    const updateData: Partial<VideoChatSession> = {};
-    if (isScheduled) {
-      updateData.scheduleTime = scheduleTime;
-    } else {
-      updateData.scheduleTime = new Date();
-    }
-
-    await this.videoChatSessionRepository.updateById(
-      sessionDetail.id,
-      updateData,
-    );
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    this.auditLogRepository.create(auditLogPayload);
+    return this.chatSessionService.editMeeting(meetingLinkId, body)
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.StopMeeting]})
+  @authorize({ permissions: [PermissionKeys.StopMeeting] })
   @patch('/session/{meetingLinkId}/end', {
     description: `Used to stop the current active meeting. Meeting cannot be stopped again if it is 
       already stopped. Successful execution will add the endTime attribute to a recently 
@@ -334,50 +132,11 @@ export class VideoChatSessionController {
   async endSession(
     @param.path.string('meetingLinkId') meetingLinkId: string,
   ): Promise<void> {
-    const auditLogPayload = {
-      action: 'session',
-      actionType: 'end-session',
-      before: {meetingLinkId},
-      actedAt: moment().format(),
-      after: {},
-    };
-    let errorMessage: string;
-    if (typeof meetingLinkId !== 'string' || !meetingLinkId) {
-      errorMessage = 'Meeting link should be a valid string.';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest('Meeting link should be a valid string.');
-    }
 
-    const videoSessionDetail = await this.videoChatSessionRepository.findOne({
-      where: {
-        meetingLink: meetingLinkId,
-      },
-    });
-
-    if (!videoSessionDetail) {
-      errorMessage = 'Meeting Not Found';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-
-    if (videoSessionDetail.endTime) {
-      errorMessage = 'Meeting has already been ended!';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.BadRequest(errorMessage);
-    }
-
-    await this.videoChatSessionRepository.updateById(videoSessionDetail.id, {
-      endTime: new Date(),
-    });
-
-    auditLogPayload.after = {response: 'end session successful'};
-    await this.auditLogRepository.create(auditLogPayload);
+    return this.chatSessionService.endSession(meetingLinkId);
   }
 
-  @authorize({permissions: ['*']})
+  @authorize({ permissions: ['*'] })
   @post('/webhooks/session', {
     description:
       'Webhook API hit from a third party to add/update session attendees in a meeting.',
@@ -390,91 +149,7 @@ export class VideoChatSessionController {
   async checkWebhookPayload(
     @requestBody() webhookPayload: VonageSessionWebhookPayload,
   ) {
-    try {
-      const {
-        connection: {data},
-        event,
-        sessionId,
-      } = webhookPayload;
-
-      const sessionAttendeeDetail =
-        await this.sessionAttendeesRepository.findOne({
-          where: {
-            sessionId: sessionId,
-            attendee: data,
-          },
-        });
-      if (!sessionAttendeeDetail) {
-        if (event === VonageEnums.SessionWebhookEvents.ConnectionCreated) {
-          await this.sessionAttendeesRepository.create({
-            sessionId: sessionId,
-            attendee: data,
-            createdOn: new Date(),
-            isDeleted: false,
-            extMetadata: {webhookPayload: webhookPayload},
-          });
-        }
-      } else {
-        const updatedAttendee = {
-          modifiedOn: new Date(),
-          isDeleted: sessionAttendeeDetail.isDeleted,
-          extMetadata: {webhookPayload: webhookPayload},
-        };
-
-        if (event === VonageEnums.SessionWebhookEvents.ConnectionCreated) {
-          updatedAttendee.isDeleted = false;
-          await this.sessionAttendeesRepository.updateById(
-            sessionAttendeeDetail.id,
-            updatedAttendee,
-          );
-        } else if (event === VonageEnums.SessionWebhookEvents.StreamCreated) {
-          await this.sessionAttendeesRepository.updateById(
-            sessionAttendeeDetail.id,
-            updatedAttendee,
-          );
-        } else if (event === VonageEnums.SessionWebhookEvents.StreamDestroyed) {
-          await this.processStreamDestroyedEvent(
-            webhookPayload,
-            sessionAttendeeDetail,
-            updatedAttendee,
-          );
-        } else if (
-          event === VonageEnums.SessionWebhookEvents.ConnectionDestroyed
-        ) {
-          updatedAttendee.isDeleted = true;
-          await this.sessionAttendeesRepository.updateById(
-            sessionAttendeeDetail.id,
-            updatedAttendee,
-          );
-        } else {
-          //DO NOTHING
-        }
-      }
-      await this.auditLogRepository.create(
-        {
-          action: 'session-webhook',
-          actionType: event,
-          before: webhookPayload,
-          after: {response: 'Webhook event triggered successfully'},
-          actedAt: moment().format(),
-        },
-        {skipCurrentUser: true},
-      );
-    } catch (error) {
-      await this.auditLogRepository.create(
-        {
-          action: 'session-webhook',
-          actionType: webhookPayload.event,
-          before: webhookPayload,
-          after: {errorStack: error.stack},
-          actedAt: moment().format(),
-        },
-        {skipCurrentUser: true},
-      );
-      throw new HttpErrors.InternalServerError(
-        'Error occured triggering webhook event',
-      );
-    }
+    return this.chatSessionService.checkWebhookPayload(webhookPayload);
   }
 
   async processStreamDestroyedEvent(
@@ -482,32 +157,18 @@ export class VideoChatSessionController {
     sessionAttendeeDetail: SessionAttendees,
     updatedAttendee: Partial<SessionAttendees>,
   ) {
-    if (
-      webhookPayload.reason === 'forceUnpublished' ||
-      webhookPayload.reason === 'mediaStopped'
-    ) {
-      await this.sessionAttendeesRepository.updateById(
-        sessionAttendeeDetail.id,
-        updatedAttendee,
-      );
-    } else {
-      updatedAttendee.isDeleted = true;
-      await this.sessionAttendeesRepository.updateById(
-        sessionAttendeeDetail.id,
-        updatedAttendee,
-      );
-    }
+   return this.chatSessionService.processStreamDestroyedEvent(webhookPayload, sessionAttendeeDetail, updatedAttendee)
   }
 
   @authenticate(STRATEGY.BEARER)
-  @authorize({permissions: [PermissionKeys.GetAttendees]})
+  @authorize({ permissions: [PermissionKeys.GetAttendees] })
   @get('/session/{meetingLinkId}/attendees', {
     security: OPERATION_SECURITY_SPEC,
-    parameters: [{name: 'active', schema: {type: 'string'}, in: 'query'}],
+    parameters: [{ name: 'active', schema: { type: 'string' }, in: 'query' }],
     responses: {
       [STATUS_CODE.OK]: {
         content: {
-          [CONTENT_TYPE.TEXT]: {schema: {type: 'array'}},
+          [CONTENT_TYPE.TEXT]: { schema: { type: 'array' } },
         },
       },
     },
@@ -516,47 +177,6 @@ export class VideoChatSessionController {
     @param.path.string('meetingLinkId') meetingLinkId: string,
     @param.query.string('active') active: string,
   ): Promise<SessionAttendees[]> {
-    const auditLogPayload = {
-      action: 'session',
-      actionType: 'session-attendees-list',
-      before: {meetingLinkId},
-      actedAt: moment().format(),
-      after: {},
-    };
-    let errorMessage: string;
-
-    const videoSessionDetail = await this.videoChatSessionRepository.findOne({
-      where: {
-        meetingLink: meetingLinkId,
-      },
-    });
-
-    if (!videoSessionDetail) {
-      errorMessage = 'Meeting Not Found';
-      auditLogPayload.after = {errorMessage};
-      await this.auditLogRepository.create(auditLogPayload);
-      throw new HttpErrors.NotFound(errorMessage);
-    }
-
-    let whereFilter = {};
-    if (active === 'true') {
-      whereFilter = {
-        sessionId: videoSessionDetail?.sessionId,
-        isDeleted: false,
-      };
-    } else {
-      whereFilter = {
-        sessionId: videoSessionDetail?.sessionId,
-      };
-    }
-
-    const sessionAttendeeList = await this.sessionAttendeesRepository.find({
-      where: whereFilter,
-    });
-
-    auditLogPayload.after = {response: 'get attendees successful'};
-    await this.auditLogRepository.create(auditLogPayload);
-
-    return sessionAttendeeList;
+    return this.chatSessionService.getAttendeesList(meetingLinkId, active);
   }
 }
